+++ h2gis-functions/src/test/java/org/h2gis/functions/BasicTest.java
@@ -32,8 +32,6 @@ import org.locationtech.jts.io.ParseException;
 import org.locationtech.jts.io.WKTReader;
 
 import java.sql.*;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Properties;
 
 import org.h2gis.utilities.GeometryTableUtilities;