+++ h2gis-functions/src/test/java/org/h2gis/functions/spatial/convert/GeometryConvertTest.java
@@ -21,13 +21,8 @@
 package org.h2gis.functions.spatial.convert;
 
 import org.h2gis.functions.factory.H2GISDBFactory;
-import org.h2gis.functions.factory.H2GISFunctions;
-import org.h2gis.functions.spatial.geometry.DummySpatialFunction;
-import org.h2gis.functions.spatial.geometry.GeometryFormsTest;
-import org.h2gis.unitTest.GeometryAsserts;
 import org.junit.jupiter.api.*;
 import org.locationtech.jts.geom.*;
-import org.locationtech.jts.geom.impl.CoordinateArraySequence;
 
 import java.sql.Connection;
 import java.sql.ResultSet;