+++ h2gis-utilities/src/test/java/org/h2gis/utilities/dbtypes/DBUtilsTest.java
@@ -19,10 +19,9 @@
  */
 package org.h2gis.utilities.dbtypes;
 
-import org.h2gis.postgis_jts_osgi.DataSourceFactoryImpl;
+import org.h2gis.postgis_jts.PostGISDBFactory;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
-import org.osgi.service.jdbc.DataSourceFactory;
 
 import javax.sql.DataSource;
 import java.io.File;
@@ -46,6 +45,7 @@ public class DBUtilsTest {
     private static Connection h2Conn;
     private static Connection postConn;
     private static Statement h2St;
+    private static final PostGISDBFactory dataSourceFactory = new PostGISDBFactory();
 
     @BeforeAll
     public static void init() throws Exception {
@@ -65,7 +65,6 @@ public class DBUtilsTest {
         props.setProperty("user", "orbisgis");
         props.setProperty("password", "orbisgis");
         props.setProperty("url", url);
-        DataSourceFactory dataSourceFactory = new DataSourceFactoryImpl();
 
         DataSource ds = dataSourceFactory.createDataSource(props);
         try {